Compare against list of current MPs
We already have an EP scraper for this, but that hasn't run for several
years, so this should pick up some newer members.

Running it is awkward, as the navigation is JS controlled, so I adapted
the original EP scraper, and ran it manually. Getting it working in a
Github Action would be much harder, but as the term is ending soon,
there hopefully won't be very many more changes before then, and it's
more useful to get the current members *now* than to spend a lot of time
adjusting this to be able to run every day.